refactor(ui): consolidate layout components into a reusable `GeneralLayout` component, streamlining structure and reducing redundancy
