+++ frontend/src/components/NavigationItem.vue
@@ -23,29 +23,11 @@
       v-if="props.icon"
       avatar
     >
-      <q-icon :name="props.icon">
-        <q-badge
-          v-if="props.preview"
-          color="secondary"
-          floating
-        />
-      </q-icon>
+      <q-icon :name="props.icon" />
     </q-item-section>
 
     <q-item-section>
       {{ props.label }}
-    </q-item-section>
-
-    <q-item-section side>
-      <q-badge
-        v-if="props.preview"
-        align="top"
-        floating
-        rounded
-        color="secondary"
-      >
-        Preview
-      </q-badge>
     </q-item-section>
   </q-item>
